Moved IAnalyzer and AnalysisLanguage to Core
