+++ src/Integration.Vsix.UnitTests/SonarLintDaemon/DaemonAnalyzerTests.cs
@@ -21,6 +21,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Integration.Vsix;
 using System;
 using System.ComponentModel;
