+++ src/Integration.Vsix.UnitTests/SonarLintDaemon/Helpers/DummySonarLintDaemon.cs
@@ -18,13 +18,12 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using EnvDTE;
-using SonarLint.VisualStudio.Integration.Vsix;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Integration.Vsix;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
