+++ src/Integration.Vsix/Analysis/IAnalyzerController.cs
@@ -18,26 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.Collections.Generic;
-using System.Threading;
 using SonarLint.VisualStudio.Core;
-
-// TODO: decide whether both of these interfaces are required
 
 namespace SonarLint.VisualStudio.Integration.Vsix.Analysis
 {
-    public interface IAnalyzer
-    {
-        bool IsAnalysisSupported(IEnumerable<AnalysisLanguage> languages);
-
-        void ExecuteAnalysis(string path,
-            string charset,
-            IEnumerable<AnalysisLanguage> detectedLanguages,
-            IIssueConsumer consumer,
-            IAnalyzerOptions analyzerOptions,
-            CancellationToken cancellationToken);
-    }
-
     // Marker interface used by for MEF exporting/importing
     // (there are multiple implementations of IAnalyzer but only one implentation
     // of this interface).
