+++ src/Integration.Vsix/SonarLintDaemon/ISonarLintDaemon.cs
@@ -19,7 +19,7 @@
  */
 
 using System;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
+using SonarLint.VisualStudio.Core;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
